Add Geoffrey's changes:
`I attach the dinghies page with the booking frame.

I also attach yachtlog.htm and thankyou-yachtlog.htm which should go in the boatlog folder.

The instructions are in the folder "booking"  and you will need to create a new file - say yacht_instructions.htm - that does not refer to the payment that is needed for motor boats.
`